blt-ossh/pyproject.toml
[project]
name = "blt-ossh"
version = "0.1.0"
description = "A magical hat that reads your GitHub profile and suggests blogs or projects you might like"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudflare-workers"
]